feat: integrate sanctuary masterworks progression
